remove duplicated VT parsing code
